requirements_gpu.txt
tensorflow-gpu
keras
h5py
Pillow
